backend/requirements_whisper.txt
# Whisper + Chatterbox Voice Assistant Dependencies
# Core dependencies from original requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.4

# Google AI and Cloud dependencies
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
google-cloud-discoveryengine==0.11.11
google-api-core==2.15.0
google-auth==2.25.2

# NEW: Whisper ASR dependencies
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
librosa>=0.9.2
soundfile>=0.12.1
ffmpeg-python>=0.2.0

# NEW: Chatterbox TTS dependencies
# Note: If chatterbox-tts is not available, we'll use fallback TTS
# chatterbox-tts>=0.1.0  # Uncomment if available

# Fallback TTS dependencies
pyttsx3>=2.90
gTTS>=2.3.0
pygame>=2.5.0  # For audio playback

# Audio processing dependencies
scipy>=1.9.0
matplotlib>=3.6.0  # For audio visualization (optional)

# Additional utilities
requests>=2.31.0
aiofiles>=23.2.1
python-multipart>=0.0.6
typing-extensions>=4.8.0 